rename theme file
